ratios refactor
