+++ src/oscillator.rs
@@ -229,7 +229,7 @@ impl Oscillator {
         StereoWaveform { l_buffer, r_buffer }
     }
 }
-//
+
 //#[cfg(test)]
 //pub mod tests {
 //    use super::*;
